# docs/swagger.yaml
swagger: '2.0'
info:
  title: Mainflux Core API
  description: Mainflux IoT platform core service API.
  version: "0.0.1"
schemes:
  - http
  - https
produces:
  - application/json
paths:
...
  /channels:
    post:
      summary: Provision new channel
      description: |
        New channel is registered. Once the channel is created,
        it is provided with an UUID (used for identifying this channel in the system).
      tags:
        - Channels
      responses:
        201:
          description: Channel is created.
          headers:
            Location:
              description: An API endpoint for accessing the channel.
              type: string
        400:
          description: Invalid channel specification submitted.
        500:
          description: Server-side error occurred.
    get:
      summary: Get all channels
      description: |
        Get all provisioned channels.
      tags:
        - Channels
      parameters:
        - name: climit
          in: query
          required: false
          type: integer
          description: Channel Limit - limits the number of retrieved channels.
        - name: elimit
          in: query
          required: false
          type: integer
          description: Entry Limit - limits the nb of retrieved channel entries.
      responses:
        201:
          description: Channels retrieved.
          schema:
            $ref: '#/definitions/Channels'
        500:
          description: Server-side error occurred.
...
definitions:
...
  Channel:
    type: object
    properties:
      id:
        type: string
        description: Id description
      name:
        type: string
        description: Name description
      description:
        type: string
        description: Description description
      owner:
        type: string
        description: Owner description
      visibility:
        type: string
        description: Visibility description
      entries:
        type: array
        items:
          properties:
            test:
              type: string
              description: Visibility description
      created:
        type: string
        description: Created description
      updated:
        type: string
        description: Updated description
  Channels:
    type: array
    items:
      $ref: '#/definitions/Channel'
...